Completely reworked cards - Joe
So, question cards are now implemented and fully automatic; we can have as many or as few as we need. All I need to do now is use the same basic code for the chance cards, and then their implementation should (emphasis on the "should") be done!

If you care about the code at all, you'll notice that the code for the cards is quite a bit different than it was in all previous versions. Thanks to input from Danny, I realised what I had previously written was surprisingly clumsy. Now, it's admittedly a little clumsy, but it's a lot cleaner than it was before.