Chance Cards and Question now work - Joe
Both cards now work as intended. Neither of them have any textures yet, and neither of them affect player movement, but the basic code is finished.

Because of how the code works, the decks that will appear in game are "dummy" decks. Basically, the code doesn't actually have the cards drawn from these decks, it just spawns the cards in place.