All 40 cards in game, with effects - Joe
All 40 cards (that's 20 chance cards and 20 question cards) are now in the game. Due to time constraints, the effects of the chance cards have been limited to either "Miss a turn" or "Move forward/backwards [x number] of spaces". Also to note is that the cards don't have any design to them - they are just white rectangles, with text.

Given more time, we would have liked to give them a specific design. And, for the sake of this project, the effects we were originally planning on giving the cards can be found on Basecamp, under the Files header.